requirements.txt
Django==1.9.6
djangorestframework==3.4.4
djangorestframework-jwt==1.8.0
django-rest-swagger==0.3.7